pyproject.toml
[tool.poetry]
name = "vr-heart-segmentation"
version = "0.1.0"
description = "Automatic segmentation of hearts with conjenital heart disease"
authors = ["Abhijit Ramesh <abhijitramesh2k@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
monai = "^1.3.0"
matplotlib = "^3.8.3"
tqdm = "^4.66.2"
nibabel = "^5.2.0"
scikit-image = "^0.22.0"
itkwidgets = "^0.32.6"
tensorboard = "^2.16.2"
kaggle = "^1.6.6"
scikit-learn = "^1.4.1.post1"
pydicom = "^2.4.4"
pynrrd = "^1.0.0"
numpy = "^1.26.4"
colorlog = "^6.8.2"
hydra-core = "^1.3.2"
lightning = "^2.2.0.post0"
wandb = "^0.16.3"
simpleitk = "^2.4.0"
streamlit = "^1.44.1"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^24.2.0"}
jupyter = "^1.0.0"
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
